refactor: 메서드 역할에 맞게 클래스 이동 #25
